requirements.txt
geopy==1.21.0
python-dotenv==0.11.0
python-telegram-bot==12.4.2
redis==3.4.1
requests==2.22.0
urllib3==1.25.8
Flask==0.12
gunicorn==19.6.0
geopy==1.21.0